+++ CHANGELOG.md
@@ -25,6 +25,10 @@ SPDX-License-Identifier: LicenseRef-GPL-2.0-only-with-DUNE-exception
   also supports the syntax `DUNE_THROW(ExceptionType, a << b) << c << d` and
   `DUNE_THROW(ExceptionType) << a << b`.
 
+- Add concepts `Std::three_way_comparable` and `Std::three_way_comparable_with` as well as an
+  algorithm `Std::lexicographical_compare_three_way` to provide library utilities for the `<=>`
+  comparison operator.
+
 - Add deduction guides to `TupleVector` analogous to `std::tuple`.
 - Add concept definition `Std::indirectly_copyable` to constrain the `DenseMatrixAssigner`.